Added __init__() method to tree assembly class
Being a little clever here by setting the list of pairwise distance frequencies
and the dict of pairwise distance histograms, both of which should be lookup
containers shared by all instances of the class, as attributes on the class.
This __init__() should be called only once, with all subsequent instances
created by copying the first instance.