TreeAssembly.__init__() accepts new kwarg absolute_freq_cutoff
This cutoff is different from constraint_freq_cutoff, which is used to
select leaf pair distances to be used as "constraints", meaning they are
used to propose new extensions to existing assemblies. That cutoff is
more stringent to help limit the number of trees being assembled
simultaneously.

On the other hand, the absolute_freq_cutoff is to be used when deciding
whether a proposed extension is "consistent" with observed distances.
Every new pairwise distance introduced by the extension must have
observed frequency above absolute_freq_cutoff, though its frequency does
not have to meet the constraint_freq_cutoff.