Switched attr unchecked_pwdists to a dict keyed by frozensets instead of tuples
The switch to dict is for faster "key in unchecked_pwdists" lookup (appears to
be ~20x faster using timeit). And I remember why I chose to use frozensets as
keys instead of tuples previously, even when their members are immutable
(making the tuples hashable and viable as keys) - the fact that
('A','B') != ('B','A') is a pain frozensets allow you to avoid.

Also attempted to prettify the nested comprehension statements by splitting
them over multiple lines and putting closing brackets/braces/parentheses on
separate lines, lined up with corresponding opening ones.