Renamed ProposedExtension.unchecked_pwdists to unverified and its changed logic
The change is somewhat subtle but important. Previously I intended this
container to hold pairs that have not yet been encountered in enumerating
constraint pairs. Then I realized that it's useless to know that an
inconsistent distance for this pair has been encountered, because a consistent
one still has to be found if the proposed extension is to be accepted. Now only
encountering a consistent constraint pair causes an unverified pair to be
popped, since in that case it becomes "verified" as having a distance
consistent with the proposed extension.