Nested in class ProposedExtension namedtuple IndexedPair for storing (i,pair)
Pairs in ProposedExtension.consistent are now stored as IndexedPair namedtuples,
still keyed by the pair of leaf names. Using the index as key, the way I do in
ProposedExtension.inconsistent would have worked too, but I find it more
convenient to use leaf name pairs as keys for consistent because it simplifies
(and speeds up) the "pair.leaves in self.consistent" check. This way indeces are
still stored.