Fixed bug brought about by caching attr _nested_set_reprs - it went out of date
When a TreeAssembly was extended with a built extension attribute
_nested_set_reprs became out of date, as it had been computed on the
pre-extension version of the object. However, since the attribute existed, its
recomputation was not triggered on the next build iteration. The problem was
propagated to copies, as the attribute is copied with everything else. I should
revisit how copying happens ... but that's for another time.

Now recomputation happens in method recompute(), which is called after extension
building. Retrieval/computing logic has been pushed out to property
current_clades_as_nested_sets, which is accessed from method as_nested_sets()