Preparing to implement speed-up of cached property recomputing
The idea is to avoid doing a full traversal of every built clade every time
an assembly is updated, the way it works now. Line-by-line profiling shows that
this accounts for substantially over half of run time.

We already compute these properties for proposed extensions during the filtering
process, but those results are thrown away afterward. Instead we'll attach them
to the extension object and, should that object be kept, use that information
when recomputing the properties of the updated assembly.

First step - pass the extension to method recompute