Added class SharedFIFOfile(FIFOfile) with synchronization-safe(ish) access
Methods read() and write() use method _sync_safe_access() to call FIFOfile's
read() and write(), respectively, and take and pass on to _sync_safe_access()
optional argument be_polite, defaulting to True, which controls how FIFOfile's
methods are called: if be_polite=True the lock is acquired just prior to
calling and released immediately after, otherwise it is neither acquired nor
released. In the latter case it must be acquired prior to the call to the
derived class method and released after.

Be careful! If called with be_polite=False without acquiring the lock first,
the calling process with deadlock! Call it the price of being impolite ...