Implemented mechanism for re-obtaining existing wrapper objects after unpickling
Right now underlying Clade objects, but not wrapper objects, are kept alive
while all the assemblies that use it are in the FIFO by maintaining a "strong"
reference to them in T._keep_alive. Number of pickled wrapper objects is tracked
and strong reference is destroyed when all wrappers are unpickled - don't want
to keep the Clade object around after all assemblies that use it have been
discarded.

This saves time on re-instantiation of the Clade objects at the cost of using
additional memory. If memory usage becomes a problem again when a very large
number of assemblies is put on the FIFO, this will have to be dispensed with.