Added to SharedFIFOfile and Event attribute and mapped its methods to the FIFO
This allows the writer process to notify the reader process that something is
available to read, and allows the reader process to sleep until something is
available by simply calling wait(). When the reader pops None, signifying that
it reached the (currently) last item in the FIFO, it clears the Event variable,
allowing it to wait again on the next call to pop(). Call to push() sets the
Event if already_have_lock is False, indicating that pop() is responsible for
FIFO state logic. Otherwise caller to pop() is responsible.