Added method start_IN_end() to initialize base FIFOfile instead of __init__()
The SharedFIFOfile object will be pickled for sending to the child queue loader
process, so it should not have open file handles at that time. Moreover, only
the parent process should have the write handle and only the child process
should have the read handle. So it makes sense to initialize the base class
object after the child process is started. On the other hand, the sync objects
(lock,event) should already exist before the child process starts. This is why
init of SharedFIFOfile and of the underlying FIFOfile were split up.