Added mechanism to pass tempfile name to FIFO belonging to child process
This is something of a Catch 22. When a NamedTemporaryFile is used, its name
is only known once its write handle is opened. At that time the child process
must already be running, so the name can't be passed to it together with the
rest of the SharedFIFOfile object. Instead it gets sent through a Pipe, which
is immediately closed afterward.

If the name is known in advance, as it is if its passed to __init__(), then
it is sent along with the SharedFIFOfile object