Workspace top-off now happens after new assemblies have had a chance to be added
This means new assemblies created during an iteration get priority for being
added to active workspace over assemblies currently in the FIFO.

This should not affect the end result when operating in single-proc mode, but
allows for using "workspace is empty and there is nothing coming to fill it up
means all assembly is finished" logic in multiproc mode. Otherwise that test
would return true even when just-created new assemblies are waiting to be added
to either the workspace or the queue.