Overrode method top_off_workspace() of base class with topping off from queue
If a situation is encountered where the workspace is empty and nothing is coming
out of the queue, even after waiting 5 seconds, it suggests that all assembly
work is complete. In that case method top_off_workspace() throws an exception
created specifically for this purpose, AssemblyWorkFinished.

Also adding class AssemblyWorkFinished(Exception)