Overrode method process_extended_assembly() of base class with multiproc way
Unlike in single proc mode, where method process_extended_assembly() of
AssemblyWorkspace sorts accepted_assemblies and updates curr_min_score, in
multiproc mode this is beyond the purview of worker processess and, therefore,
of process_extended_assembly() methods of their WorkerProcAssemblyWorkspace
objects. Instead scores of completed assemblies that are better than the
current min score (a shared Value which will be updated by the main process)
are put on the scores queue and the opportunity is taken to filter this
workspace's accepted_assemblies based on curr_min_score. However, there is no
guarantee that curr_min_score has already been updated by the main process to
reflect the just-submitted score. In fact, it probably hasn't been, but that
should be OK - a future round of filtering will take care of it.