Added bare framework for what main process does to manage multiproc assembly
Shared assembly queue and encountered_assemblies_dict objects, a seed_assembly,
and a scores_queue are instantiated. Worker processes are run in a try block
(their instantiation is not yet implemented) and terminated in the
corresponding except block if something goes wrong. Managers for assembly queue
and encountered_assemblies_dict shared objects are then shut down in a finally
block. A results object, whatever it may end up being, is returned if workers
finish properly.