Fixed bug where all workers were getting the same seed assembly - the empty one
Assembly procedure is now "primed" until there are more seed assemblies than the
requested number of workers. Each worker then gets a unique seed assembly and
any remaining ones are put into the assembly queue (prior to which they are
pickled).

Unfortunately this means the seed assembly can't be packaged with the rest of
the arguments for workspace object instantiation, so passing of arguments to
AssemblerProcess instantiaion is also affected