Fixed bug where all worker processes got the same seed assembly
Until workers are kicked off by calling their run() methods, they exist in the
memory space of their parent process, meaning they all get the same ArgsKwargs
namedtuple object. When it was modified in the __init__() of one process obj,
even after it was named self.pass_to_workspace, it actually got modified for
all processes, so that in the end all processes got the same seed assembly -
the one that was passed to the last process.

This resulted in missing some final trees that were unreacheable without
starting with the missing seed