Started work on FIFOfile implementation with temp file rollover
All temp files for each FIFO will be located in a temp directory. To this end,
adding module tempdir, an exact lift of the module with same name from IPython,
which itself is a backport of tempdir functionality added to the standard lib in
python 3.2. We want the TemporaryDirectory context manager implemented there.