Fixed small bugs: spurious parameter self to __new__() and spelling mistake
Also moved creation+initialization of class parameter TMPFILE.instcount to
classmethod init_class().

Unfortunately test continues to fail, but now it does so because of attempting
to pass a Mock object (self.name) to open() instead of a string when opening
the writing handle, wh. Will fix this when implementing the writing interface
of TMPFILE.