Modified FIFOfile instantiation and added a test of same
Removed redundant call to tmpdir_obj.__enter__(), which was a bug. Also removed
instatiation of first TMPFILE object, now to be done in a separate method, and
cleartion of attribute file_spool, which is now a class attribute of class
FIFOfile.TMPFILE, rather than an attribute of FIFOfile instances. This last is
an effort to keep all interactions with real files, to be handled by TMPFILE,
separate from the FIFO interface, to be handled by FIFOfile.