Fixed bug in testing code that caused some tests to fail spuriously
Because class FIFOfile, and therefore the class it contains, TMPFILE, were
defined once when module topolenum was imported into the testing module,
modifications made to TMPFILE in the course of one test remained when later
tests were run. Reloading the module in setUp() of each test fixture fixes the
problem, since each test method runs in a separate instance of the fixture.