Modified property rh to check whether current file is at EOF and rollover
The reading handle can be in one of two states: either at the beginning of a
pickle string or at EOF, characterized by the fact that calling read() on the
handle does not advance the cursor. If read() advances the cursor, then there is
a pickle to be read, so the cursor is returned to its starting position and the
handle is returned. If read() fails to advance the cursor, then the current file
is at EOF. This can happen if the reading cursor has caught up to the writing
cursor, in which case it will again be possible to read from the reading cursor
once the writing cursor writes a pickle. On the other hand, if another TMPFILE
object is waiting in the file_spool, then the writing handle has already rolled
over at least once, and the current file must be at its ultimate EOF. In that
case the current file is discarded (the underlying NamedTemporaryFile closed)
and replaced with the new one popped from the spool.

Also added test of new functionality