Modified logic or cursor repositioning after reading handle retrieval
Previously the cursor was repositioned to the return value of first call to
tell() if the call to read() advanced it, meaning there is something to read.
However, as the newly added comment explains, at least on OSX (but not on
Ubuntu) calling read() on a handle that is already at EOF puts the handle in a
permanently non-readable state. Even after more is written to the file starting
at the position reported by tell(), calls to read() return an empty string and
do not advnace the cursor. Repositioning the cursor to the position tell() -
handle.seek(handle.tell()) - fixes the problem. This means that in the case
when handle is at EOF but there is no spooled file to replace the current one
the cursor needs to be repositioned exactly as if the handle had something to
read.

Previously this would have been handled by pop(), but that unnecessarily
required both property rh and pop() to worry about cursor positioning. By
repositioning the cursor regardless of outcome we relieve pop() of all concerns
about the cursor. Reassigning old_pos in case of file rollover ensures the same
call, seek(old_pos), produces the result appropriate for that case.

Updated test accordingly and added explanatory comments