Added class SharedFIFOfile.SpoolerThread(threading.Thread)
This represents a separate thread of the file reading process which monitors
and responds to requests made by the file writing process for spooling a new
temp file. The thread can be shut down by calling the set() method of its
attribute stop, a threading.Event object. Thread objects are not attached to
the TMPFILE class and are not aware of it - they will be launched by and
attached to SharedeFIFOfile objects directly. Instead they require init args
connection and spool_callable, corresponding, respectively,  to the file
reading proc's end of the pipe belonging to the class TMPFILE and the
classmethod spool() of SharedFIFOfile's base class, FIFOfile. In response to
receiving a message through the connection, a request, the thread calls spool()
and sends the name of the returned TMPFILE object through the connection.

Also added test for class SpoolerThread