Updated method SharedFIFOfile.start_OUT_end() to use TMPFILE and SpoolerThread
Creation of first temp file and startup of spooling by TMPFILE is delegated to
method start_OUT_end() of base class FIFOfile. Name of newly created temp file
is sent to the writing end of the FIFO, as the startup of that end must open a
writing handle to this file. A spooler thread is initialized with the conn obj
and classmethod spool() of base class FIFOfile which will be used to spool new
temp files upon request from the writing end and send their names back to the
writing end for opening corresponding writing handles. The thread is then
started up.

Also added test of FIFO reading end startup