Updated method SharedFIFOfile.start_IN_end() to use TMPFILE and SpoolerThread
The name of the first temp file is received from the reading end, meaning the
call blocks until method start_OUT_end(), called by the process which holds the
FIFO's reading end, creates the temp file and sends its name. The name is used
to initialize a TMPFILE object on the writing end, which is then opened for
writing and set as the current_writing_file.

Also added test of FIFO writing end startup