Moved instantiation of SharedFIFOfile from __init__() of worker proc to run()
Instantiation of the SharedFIFOfile object triggers creation of connection
objects attached to class SharedFIFOfile.TMPFILE. Because during calls to their
__init__() methods worker processes still live in the memory space of the parent
process, instantiation of the FIFO object during this call causes all FIFOs to
hold the same pair of connections. Now, by the time the FIFOs are instantiated
during the run of the worker proc, each proc lives in its own memory space and
the objects representing SharedFIFOfile.TMPFILE classes in their memory spaces
are distinct, as are the created connection objects.