Updated SharedFIFOfile.close() with new implementation of closing procedure
Closing must be done in specific order. First, the writing side must close its
writing handle. Then the reading side must be allowed to carry out any remaining
reading activities and close/discard temp file(s). Finally, the writing side
can exit the temp dir context by calling tmpdir_obj.__exit__(), thereby deleting
the temp directory. After that it closes the connections attached to the TMPFILE
class.

This shutdown order assumes that the writing end is held by the parent process,
which is the case for the topology enumeration procedure, but not in the general
case. Currently FIFOs are also used for passing accepted assemblies by worker
processes back to their parent process at the end of assembly. In that case the
reading end is held by the parent process and writing ends are held by child
processes. If FIFOs continue to be used for this purpose (Pipes or Queues may be
more appropriate), or in case of some future application, the implementation
will need to be modified to be more flexible with respect to process "lineage"
and FIFO end ownership.

Also added testing of FIFO closing from POVs of procs holding reading and
writing ends