Top-level enumeration control handled in MainTopologyEnumerationProcess process
The parent process of this runner process - the main process - is now free to
observe the enumeration process and, if necessary, interrupt it with a proper
shutdown. Resources shared by the top-level runner and workers (assembly queue,
scores queue, min score value, etc) are instantiated during init of the
MainTopologyEnumerationProcess process. This means they exist in the memory
space of the main process and are accessible as attributes of the
MainTopologyEnumerationProcess object. This fact is exploited in passing
results from workers via the results queue, with the main process retrieving
results directly from the workers, bypassing the top-level runner. However, it
also means that these resources cannot be freed during the run of the top level
process. Because of this MainTopologyEnumerationProcess has a clean_up() method
which is called after that process is joined or terminated.