Fixed bug in testing code that caused test failure - file list now being sorted
It appears there are no guarantees about the order in which file names appear
in the list returned by os.listdir() - they may not be in alphabetical order.
The test relies on file name starting with FIFOfile001 appearing at index 0,
before file name starting with FIFOfile002. To guarantee this we sort the list
before performing the test.