First assembly is now instantiated during init of top-level enumeration process
The call to TreeAssembly.__init__() attaches attribute to the TreeAssembly class
itself, and now those exist in the memory space of the main process. This fixes
a bug where receiving completed assemblies (which requires pickling/unpickling)
from workers failed because the TreeAssembly class in the main process' memory
space did not have the leaves_master class attribute now required to unpickle
TreeAssembly objects.