AssemblerProcess now signals shutdown to its QueueLoader via Event close_fifo
Previous way of signaling - pushing 'SHUTDOWN' into FIFO - relies on loader
actually popping the signal. When loader trails far behind the writing handle,
catching up can take a substantial amount of time. Moreover, if the queue is
full, the loader repeatedly attempts to put the same item until it succeeds.
In the worst case, the top-level runner proc times out waiting to join each
worker and kills the worker while it's waiting to shut down its FIFO, because
it is waiting on its loader to put something on the full queue. This results in
failure to clean up the temp dir.

Now the signal is passed from worker to loader via a Event independent of the
queue. When the Event (close_fifo) is set the loader abandons attempts to put
an item if it's trying to do so, exits the while loop controlling its main
activity, closes its end of the FIFO, and exits.