Switched attr reached_num_requested_trees from bool flag to computed property
The mechanism of making this check is different in single proc and multiproc
modes, but I would like to re-use in both modes code that makes the check.
Encapsulating the check mechanism allows overriding the base class'
implementation without breaking that code.

Also dropped one-time sorting of workspace on reaching requested number of
topologies: logic of its implementation became more difficult with the swtich
to computed property, and it was (almost) never done in multiproc operation
anyway