Fixed bug introduced by the previous bugfix where workers hung during topoff
In particular this happened at the very end, when the queue became empty. The
multiproc version of fill_workspace_from_fifo() returns if it fails to obtain an
assembly from the queue and the workspace is not empty. Otherwise it tries again
and raises AssemblyWorkFinished on second failure. However, since the call to
fill_workspace_from_fifo() is now itself in a while loop in top_off_workspace(),
the worker proc could stay in this outer while loop forever when the queue is
empty. Now the loop is exited if rejected topoff assembly counter has not
reached 100, as the only other reason fill_workspace_from_fifo() returns is due
to queue being empty.