Finally fixed tacit bug where SharedFIFOfile's reading handles were not closed
The process responsible for reading from the FIFO would wait indefinitely for
the writing process to set the FIFO's event variable, indicating that it had put
something in the FIFO. When the writing process went to shut down the FIFO, the
reading process was still waiting. The writing process would timeout its wait on
the reading process and exit the tmpdir_obj context. Normally this would remove
the directory and everything in it, despite reading handles still being open, so
everything would look normal. On NFS systems the temp directory could not be
removed because, after the TemporaryDirectory mechanism cleaned it out, NFS
would create .nfs files for each removed file with an open handle.