Added computed properties nodes_left_to_build and built_nodes_count
These report the number of internal nodes left to build in order to complete
the tree and the number of nodes built so far, respectively. Former is computed
according to number internal nodes = number of leaves - 1, where number of
leaves is the sum of free leaves and built clades. Latter is computed from
former and new class attribute, total_nodes_to_build, which is itself computed
as number of leaves to assemble - 1.