Dropped CachedNewAssembly event in favor of Extended event
On every assembly extension one of the possible extensions is built by modifying
in place the assembly under consideration. This extension would get recorded as
a new assembly. This caused some debugging annoyances. An Extended event will be
reported for these as well, and by dropping the CachedNewAssembly event
redundancy can still be avoided.