Enqueuing extracted to method so it can be used with both normal and early term
This change exposed race conditions between save writing and terminating the
enumeration process after failing to join it for 30 seconds. Typically the
process gets terminated before save is written. 30 second opportunity to join
followed by termination was replaced with indefinite wait to join.
Unfortunately this creates the risk of turning into a zombie, but that will
have to be addressed separately later.