First implementation of best_case_with_extension()
This implementation is almost a word-for-word port of the original look_ahead(),
with modifications made to insert it into the new framework. It's very wasteful,
since it would result in recomputing distances to root for leaves already
attached once for every proposed extension for which the method is called.
These should be chached by the TreeAssembly obj.