Dict new_pairs now keyed by index of new pair in constraints_idx, not leaves
I made this change in order to make popping extensions from their containers
uniform, and that is the case regardless of what is used as the key. Choice of
frozenset(leaves) was arbitrary. But indeces of new pairs should actually be
useful for popping them from constraints_idx when the new pair extensions are
built. Indeces of consistent and inconsistent pairs are stored in
ProposedExtension objects for this reason.