Fixed what I believe to be a logic bug, though it's possible I introduced one
This bug SHOULDN'T have affected the outcome, as it at simply caused one of the
filters not to work, which should have simply resulted in tree assemblies
sticking around longer than necessary, which should have only had an effect on
running time, not outcome.

Need to re-run code on data where expected output is known, or at least already
obtained with the previous version of this code.