Fixed bug where extensions were added without any consistent constraint pairs
This would happen when an inconsistent constraint pair triggered creation of the
extension object, all subsequent constraint pairs were also inconsistent,
leaving all of the extension's pairs unverified, and then the first verification
filter verified all of the extension's pairs in the
freq_cutoff > pair.freq > abs_cutoff range. Specifically additions of a new leaf
to a clade with two leaves, with only two pair dists to verify, are susceptible
to this bug.