Major changes in how the workspace is refilled between iterations
Central idea: number of nodes remaining to finish in an assembly is inversely
related to the ratio of assemblies pushed to FIFOs / assemblies loaded from
FIFOs back into the workspace. When the ratio is skewed toward pushed assemblies
- i.e. there are lots of aseemblies in FIFOs - only those assemblies closest to
completion are accepted for workspace topoff. When the ratio is skewed in the
opposite direction - i.e. FIFOs are nearly or completely empty - assemblies with
any amount of work remaining to be done are accepted.

The overall effect is to prevent FIFOs from growing extremely large due to the
combinatorial explosion of assembly paths